# Start on #159: Make opening and closing transactions faster