--- a/src/relstorage/adapters/poller.py
+++ b/src/relstorage/adapters/poller.py
@@ -15,12 +15,11 @@ from __future__ import absolute_import, print_function
 
 import logging
 
-from ZODB.POSException import ReadConflictError
-from ZODB.POSException import Unsupported
 from zope.interface import implementer
 
 from ._util import formatted_query_property
 from .interfaces import IPoller
+from .interfaces import StaleConnectionError
 
 log = logging.getLogger(__name__)
 
@@ -67,13 +66,6 @@ class Poller(object):
     _poll_inv_exc_query = formatted_query_property('_poll_inv',
                                                    extension=' AND tid != %(self_tid)s')
 
-    _tran_exists_queries = (
-        "SELECT 1 FROM transaction WHERE tid = %(tid)s",
-        Unsupported("Transaction data not available without history")
-    )
-
-    _tran_exists_query = formatted_query_property('_tran_exists')
-
 
     def __init__(self, poll_query, keep_history, runner, revert_when_stale):
         self.poll_query = poll_query
@@ -85,17 +77,18 @@ class Poller(object):
         """
         Polls for new transactions.
 
-        conn and cursor must have been created previously by
-        open_for_load(). prev_polled_tid is the tid returned at the
-        last poll, or None if this is the first poll. If ignore_tid is
-        not None, changes committed in that transaction will not be
-        included in the list of changed OIDs.
+        *conn* and *cursor* must have been created previously by
+        ``open_for_load()`` (a snapshot connection). prev_polled_tid
+        is the tid returned at the last poll, or None if this is the
+        first poll. If ignore_tid is not None, changes committed in
+        that transaction will not be included in the list of changed
+        OIDs.
 
-        Returns (changes, new_polled_tid), where changes is either a
-        list of (oid, tid) that have changed, or None to indicate that
-        the changes are too complex to list --- this must cause local
-        storage caches to be invalidated.. new_polled_tid can be 0 if
-        there is no data in the database.
+        Returns ``(changes, new_polled_tid)``, where *changes* is
+        either a list of ``(oid_int, tid_int)`` that have changed, or
+        ``None`` to indicate that the changes are too complex to li
+        --- this must cause local storage caches to be invalidated..
+        *new_polled_tid* can be 0 if there is no data in the database.
         """
         # pylint:disable=unused-argument
         # find out the tid of the most recent transaction.
@@ -106,6 +99,7 @@ class Poller(object):
             # the root object.
             # Anything we had cached is now definitely invalid.
             return None, 0
+
         new_polled_tid = rows[0][0]
         if prev_polled_tid is None:
             # This is the first time the connection has polled.
@@ -138,30 +132,24 @@ class Poller(object):
             # raise ReadConflictError to trigger a retry.
             # We're probably just waiting for async replication
             # to catch up, so retrying could do the trick.
-            raise ReadConflictError(
-                "The database connection is stale: new_polled_tid=%d, "
-                "prev_polled_tid=%d." % (new_polled_tid, prev_polled_tid))
-
+            raise StaleConnectionError.from_prev_and_new_tid(
+                prev_polled_tid, new_polled_tid)
 
         # New transaction(s) have been added.
-        if self.keep_history:
-            # If the previously polled transaction no longer exists,
-            # the cache is too old and needs to be cleared.
-            # XXX Do we actually need to detect this condition? I think
-            # if we delete this block of code, all the unreachable
-            # objects will be garbage collected anyway. So, as a test,
-            # there is no equivalent of this block of code for
-            # history-free storage. If something goes wrong, then we'll
-            # know there's some other edge condition we have to account
-            # for.
-            cursor.execute(
-                self._tran_exists_query,
-                {'tid': prev_polled_tid})
-            rows = cursor.fetchall()
-            if not rows:
-                # Transaction not found; perhaps it has been packed.
-                # The connection cache should be cleared.
-                return None, new_polled_tid
+
+        # In the past, but only for history-preserving databases, we
+        # would check to see if the previously polled transaction no
+        # longer exists in the transaction table. If it didn't, we
+        # would return ``(None, new_polled_tid)``, in order to clear
+        # the Connection cache.
+        #
+        # However, we ran for yers without an analogous case for
+        # history-free databases without problems, on the theory that
+        # all the unreachable objects will be garbage collected
+        # anyway.
+        #
+        # Thus we became convinced it was safe to remove the check in history-preserving
+        # databases.
 
         # Get the list of changed OIDs and return it.
         stmt = self._poll_inv_query