--- a/src/relstorage/storage/connections.py
+++ b/src/relstorage/storage/connections.py
@@ -23,83 +23,90 @@ from __future__ import print_function
 
 import contextlib
 
+from zope.interface import implementer
+
+from .._util import Lazy
+from .interfaces import IManagedDBConnection
+
 logger = __import__('logging').getLogger(__name__)
 
-class AbstractConnection(object):
-    """
-    Represents a database connection and its single cursor.
+__all__ = [
+    'LoadConnection',
+    'StoreConnection',
+]
 
-    If the connection is not open and presumed to be good,
-    this object has a false value.
 
-    "Restarting" a connection means to bring it to a current view of the
-    database. Typically this means a rollback so that a new transaction
-    can begin with a new MVCC snapshot.
-    """
+class AbstractManagedConnection(object):
 
-    __slots__ = (
-        'connection',
-        'cursor',
-        'connmanager',
-    )
+    _NEW_CONNECTION_NAME = None
+    _RESTART_NAME = None
 
     def __init__(self, connmanager):
         self.connection = None
-        self.cursor = None
         self.connmanager = connmanager
+        self._cursor = None
+        self.active = False
+        self._new_connection = getattr(connmanager, self._NEW_CONNECTION_NAME)
+        self._restart = getattr(connmanager, self._RESTART_NAME)
+
+    # Hook functions
+    on_opened = staticmethod(lambda conn, cursor: None)
+    on_rolledback = on_opened
+    on_first_use = on_opened
 
     def __bool__(self):
-        return self.connection is not None and self.cursor is not None
+        return self.connection is not None and self._cursor is not None
 
     __nonzero__ = __bool__
 
+    @Lazy
+    def cursor(self):
+        if not self.active or not self:
+            # XXX: If we've been explicitly dropped, do we always want to
+            # automatically re-open? Probably not; bad things could happen:
+            # a load connection could skip into the future, and a store connection
+            # could lose temp tables.
+            conn, cursor = self.open_if_needed()
+            self.active = True
+            self.on_first_use(conn, cursor)
+        return self._cursor
+
     def drop(self):
-        """
-        Unconditionally drop the connection.
-        """
-        conn, cursor = self.connection, self.cursor
-        self.connection, self.cursor = None, None
+        self.active = False
+        conn, cursor = self.connection, self._cursor
+        self.connection, self._cursor = None, None
+        self.__dict__.pop('cursor', None)
         self.connmanager.rollback_and_close(conn, cursor)
 
     def rollback(self):
-        """
-        Rollback the connection.
-
-        When this completes, the connection will be in a neutral state,
-        not idle in a transaction.
-
-        If an error occurs during rollback, the connection is dropped.
-        """
+        self.active = False
         if not self:
             return
 
+        conn = self.connection
+        cur = self._cursor
+        self.__dict__.pop('cursor', None)
         try:
-            self.connmanager.rollback(self.connection, self.cursor)
+            self.connmanager.rollback(conn, cur)
         except:
             self.drop()
             raise
+        finally:
+            self.on_rolledback(conn, cur)
 
     def open_if_needed(self):
         if not self:
             self.drop()
             self._open_connection()
-        return self.connection, self.cursor
-
-    def _restart(self):
-        raise NotImplementedError
+        return self.connection, self._cursor
 
     def _open_connection(self):
         """
         Open a new connection, assigning it to ``connection`` and ``cursor``
         """
         new_conn, new_cursor = self._new_connection()
-        self.connection, self.cursor = new_conn, new_cursor
-
-    def _new_connection(self):
-        """
-        Open and return a new ``(connection, cursor)`` pair.
-        """
-        raise NotImplementedError
+        self.connection, self._cursor = new_conn, new_cursor
+        self.on_opened(new_conn, new_cursor)
 
     def __noop(self, *args):
         "does nothing"
@@ -108,21 +115,28 @@ class AbstractConnection(object):
         """
         Unconditionally restart the connection.
         """
+        self.active = False
         self.restart_and_call(self.__noop)
 
     def restart_and_call(self, f, *args, **kw):
         """
-        Restart the connection and call a function. This may
-        drop and reload the connection if necessary.
+        Restart the connection (roll it back) and call a function
+        after doing this.
 
-        :param callable f: The function to call: ``f(conn, cursor, *args, **kwargs)``.
-            May be called up to twice.
+        This may drop and re-connect the connection if necessary.
+
+        :param callable f:
+            The function to call: ``f(conn, cursor, *args, **kwargs)``.
+            May be called up to twice if it raises a disconnected exception
+            on the first try.
+
         :return: The return value of ``f``.
         """
         def callback(conn, cursor, fresh, *args, **kwargs):
+            assert conn is self.connection and cursor is self._cursor
             if not fresh:
-                assert conn is self.connection and cursor is self.cursor
-                self._restart()
+                self._restart(conn, cursor)
+                self.on_rolledback(conn, cursor)
             return f(conn, cursor, *args, **kwargs)
 
         return self.call(callback, True, *args, **kw)
@@ -132,17 +146,21 @@ class AbstractConnection(object):
         Call a function with the cursor, connecting it if needed.
         If a connection is already open, use that without rolling it back.
 
+        Note that this does not count as a first usage and won't invoke
+        that callback.
+
         :param callable f: Function to call
-            ``f(store_conn, store_cursor, fresh_connection, *args, **kwargs)``.
-            The function may be called up to twice, if the *fresh_connection* is false
+            ``f(conn, cursor, fresh_connection_p, *args, **kwargs)``.
+            The function may be called up to twice, if the *fresh_connection_p* is false
             on the first call and a disconnected exception is raised.
         :keyword bool can_reconnect: If True, then we will attempt to reconnect
-            the connection and try again if an exception is raised. If False,
+            the connection and try again if an exception is raised if *f*. If False,
             we let that exception propagate. For example, if a transaction is in progress,
             set this to false.
         """
         fresh_connection = False
         if not self:
+            # We're closed or disconnected. Start a new connection entirely.
             self.drop()
             self._open_connection()
             fresh_connection = True
@@ -150,7 +168,7 @@ class AbstractConnection(object):
             can_reconnect = False
 
         try:
-            return f(self.connection, self.cursor, fresh_connection, *args, **kwargs)
+            return f(self.connection, self._cursor, fresh_connection, *args, **kwargs)
         except self.connmanager.disconnected_exceptions as e:
             if not can_reconnect:
                 raise
@@ -162,116 +180,56 @@ class AbstractConnection(object):
                 logger.exception("Reconnect %s failed", self)
                 raise
             logger.info("Reconnected %s", self)
-            return f(self.connection, self.cursor, True, *args, **kwargs)
+            return f(self.connection, self._cursor, True, *args, **kwargs)
 
     @contextlib.contextmanager
     def isolated_connection(self):
-        """
-        Context manager that opens a distinct connection and return
-        its cursor.
-
-        No matter what happens in the block, the connection will be
-        dropped afterwards.
-        """
         conn, cursor = self._new_connection()
         try:
             yield cursor
         finally:
             self.connmanager.rollback_and_close(conn, cursor)
 
-class LoadConnection(AbstractConnection):
+
+@implementer(IManagedDBConnection)
+class LoadConnection(AbstractManagedConnection):
 
     __slots__ = ()
 
-    def _new_connection(self):
-        return self.connmanager.open_for_load()
-
-    def _restart(self):
-        self.connmanager.restart_load(self.connection, self.cursor)
+    _NEW_CONNECTION_NAME = 'open_for_load'
+    _RESTART_NAME = 'restart_load'
 
 
-class StoreConnection(AbstractConnection):
+@implementer(IManagedDBConnection)
+class StoreConnection(AbstractManagedConnection):
 
     __slots__ = ()
 
-    def _new_connection(self):
-        return self.connmanager.open_for_store()
-
-    def _restart(self):
-        self.connmanager.restart_store(self.connection, self.cursor)
+    _NEW_CONNECTION_NAME = 'open_for_store'
+    _RESTART_NAME = 'restart_store'
 
 
-class EventedConnectionWrapper(object):
-    """
-    A facade around a connection that keeps track of certain
-    events about the state of the connection and
-    handles those events automatically.
-
-    Events are handled via callbacks on this object. You can subclass
-    or you can assign to them.
-    """
-
-    # Assigning to hook attributes makes it more difficult to use
-    # __slots__ because the method and slot conflict.
-
-    def __init__(self, connection):
-        """
-        The connection should not already be open.
-        """
-        assert not connection
-        self.__connection = connection
-        self.__active = False
-        self.isolated_connection = connection.isolated_connection
-        self.restart_and_call = connection.restart_and_call
-
-    def __bool__(self):
-        return bool(self.__connection)
-
-    __nonzero__ = __bool__
-
-    @property
-    def connection(self):
-        return self.__connection.connection
-
-    @property
-    def cursor(self):
-        if not self.__active or not self.__connection:
-            self.__active = True
-            conn, cursor = self.__connection.open_if_needed()
-            self.activated(conn, cursor)
-        return self.__connection.cursor
-
-    def activated(self, connection, cursor):
-        """
-        Hook method you can assign to notice when a cursor becomes active.
-        """
-
-    def drop(self):
-        self.__active = False
-        self.__connection.drop()
-
-    def rolledback(self):
-        """
-        Hook method. Inspect sys.exc_info() to determine if we're called with an
-        exception.
-
-        Always called, even in the event of a disconnection.
-        """
-
-    def rollback(self):
-        self.__active = False
-        try:
-            self.__connection.rollback()
-        finally:
-            self.rolledback()
-
+@implementer(IManagedDBConnection)
 class ClosedConnection(object):
     """
     Represents a permanently closed connection.
     """
     __slots__ = ()
 
+    cursor = None
+    connection = None
+
+    def __init__(self, *args):
+        "We have no state."
+
     def drop(self):
         "Does nothing."
 
     rollback = drop
+
+    __bool__ = __nonzero__ = lambda self: False
+
+    def isolated_connection(self, *args, **kwargs):
+        raise NotImplementedError
+
+    restart_and_call = isolated_connection