… CHANGES.rst
… src/relstorage/storage/interfaces.py
… src/relstorage/storage/tests/test_connections.py
… src/relstorage/tests/reltestbase.py