This is the beginning of some work on #159. We start here by eliminating some unnecessary database traffic that we can control.

`zodbshootout --loops 32 --values 4   -c 5 -p 10`

| Benchmark                             | master | This branch               |
|---------------------------------------|-------------|-----------------------------|
| MappingStorage: empty implicit commit | 40.9 us     | 44.0 us: 1.08x slower (+8%) |
| psycopg2_hp: empty implicit commit    | 435 us      | 424 us: 1.03x faster (-3%)  |
| mysqlclient_hp: empty explicit commit | 771 us      | 752 us: 1.03x faster (-3%)  |
| mysqlclient_hp: empty implicit commit | 828 us      | 624 us: 1.33x faster (-25%) |
| psycopg2_hf: empty explicit commit    | 538 us      | 449 us: 1.20x faster (-17%) |
| psycopg2_hf: empty implicit commit    | 553 us      | 425 us: 1.30x faster (-23%) |
| mysqlclient_hf: empty explicit commit | 912 us      | 764 us: 1.19x faster (-16%) |
| mysqlclient_hf: empty implicit commit | 891 us      | 644 us: 1.38x faster (-28%) |

Not significant (2): MappingStorage: empty explicit commit; psycopg2_hp: empty explicit commit